fix(NodeAllocator): fix network_id when using weave and prevent removing the network

+++ src/hosted/NodeAllocator.cc
@@ -181,9 +181,15 @@ ebbrt::NodeAllocator::NodeAllocator() : node_index_(2), allocation_index_(0) {
                         ".IPAM.Config}}{{.Gateway}}{{end}}' " +
                         network_id_);
   } else {
-    std::cerr << "Creating custom network" << std::endl;
-    network_id_ = RunCmd(CustomNetworkCreate);
-    network_ip = RunCmd(CustomNetworkIp + network_id_);
+    if (CustomNetworkNodeArguments.find("weave") != std::string::npos) {
+      std::cerr << "Connecting to weave network" << std::endl;
+      network_id_ = "weave";
+      network_ip = RunCmd(CustomNetworkIp);
+    } else {
+      std::cerr << "Creating custom network" << std::endl;
+      network_id_ = RunCmd(CustomNetworkCreate);
+      network_ip = RunCmd(CustomNetworkIp + network_id_);
+    }
   }
 
   uint8_t ip0, ip1, ip2, ip3;
@@ -343,6 +349,10 @@ ebbrt::NodeAllocator::AllocateNode(std::string binary_path, int cpus,
 }
 ebbrt::NodeAllocator::~NodeAllocator() {
   nodes_.clear();
+  /* If weave is the network don't remove it */
+  if (CustomNetworkNodeArguments.find("weave") != std::string::npos) {
+    return;
+  }
   if (CustomNetworkRemove.empty()) {
     std::cerr << "removing Network: " << network_id_.substr(0, 12) << std::endl;
     RunCmd("docker network rm " + network_id_);
